requirements.txt
google-genai
fastapi
uvicorn[standard]
asyncpg
